Improve StackItem previews
