+++ previews/primer/alpha/stack_item_preview.rb
@@ -4,13 +4,33 @@ module Primer
   module Alpha
     # @label StackItem
     class StackItemPreview < ViewComponent::Preview
+      # @label Default
+      def default
+      end
+
       # @label Playground
       #
+      # @param tag text
       # @param grow toggle
-      def playground(grow: false)
+      # @param grow_narrow toggle
+      # @param grow_regular toggle
+      # @param grow_wide toggle
+      def playground(
+        tag: Primer::Alpha::StackItem::DEFAULT_TAG,
+        grow: Primer::Alpha::StackItem::GrowArg::DEFAULT,
+        grow_narrow: Primer::Alpha::StackItem::GrowArg::DEFAULT,
+        grow_regular: Primer::Alpha::StackItem::GrowArg::DEFAULT,
+        grow_wide: Primer::Alpha::StackItem::GrowArg::DEFAULT
+      )
         render_with_template(locals: {
           system_arguments: {
-            grow: grow
+            grow: [
+              grow,
+              grow_narrow,
+              grow_regular,
+              grow_wide,
+              grow_wide
+            ]
           }
         })
       end
