+++ previews/primer/alpha/stack_item_preview/default.html.erb
@@ -1,0 +1,5 @@
+<div class="d-flex border color-border-success-emphasis">
+  <%= render(Primer::Alpha::StackItem.new(border: true, border_color: :danger_emphasis)) do %>
+    Hello, world!
+  <% end %>
+</div>
